Save all 64 bits of get_ticks_msec() in more cases
(cherry picked from commit 5dc02eb8b0e30a3ede3be7e341a2681e92cf3562)

+++ core/bind/core_bind.cpp
@@ -919,7 +919,7 @@ void _OS::delay_msec(int p_msec) const {
 	OS::get_singleton()->delay_usec(int64_t(p_msec) * 1000);
 }
 
-uint32_t _OS::get_ticks_msec() const {
+uint64_t _OS::get_ticks_msec() const {
 	return OS::get_singleton()->get_ticks_msec();
 }
